requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
alembic
email-validator
python-multipart
python-dotenv
groq
fpdf2
redis
argon2_cffi
jinja2
weasyprint
PyMuPDF
pdf2image 
pillow